backend/pyproject.toml
[tool.poetry]
name = "brainwave-insightgpt"
version = "0.1.0"
description = ""
authors = ["Priyanshu <priyanshunegi246@gmail.com>"]
readme = "README.md"
packages = [{include = "backend", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.128.0"
pydantic = "^2.5.0"
uvicorn = "^0.40.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
httpx = "^0.28.1"
aiofiles = "^23.2.1"
textstat = "^0.7.3"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
google-cloud-firestore = "^2.23.0"
google-cloud-documentai = "^3.6.0"
google-cloud-storage = "^2.12.0"
google-cloud-dlp = "^3.14.0"
google-cloud-pubsub = "^2.19.0"
google-cloud-bigquery = "^3.13.0"
google-cloud-translate = "^3.15.0"
google-auth = "^2.25.2"
google-genai = "1.33.0"
google-generativeai = "^0.8.5"
pandas = "^2.1.4"
numpy = "2.3.2"
scikit-learn = "^1.3.2"
regex = "2025.9.1"
PyPDF2 = "^3.0.1"
pdfminer-six = "^20231228"
requests = "^2.32.5"
email-validator = "^2.3.0"

[tool.poetry.scripts]
start = "backend.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
